Add redis failover solution
Due to the fact that redis has no autom. failover mechanism, it is now
possible to switch on 'db_failover' config option, so that 'ab_test' and
'finished' will not crash. 'ab_test' always delivers alternative A (the
first one) in that case. It's also possible to set a 'db_failover_on_db_error'
handler (proc) in the config, for example to log these errors via
Rails.logger.