Fixes version number wackiness when alternatives are changed
I found that when I updated the alternatives in a test it kept resetting. I must have not noticed for a while, because when I checked, I was on version 30,561 yet had no results.
It turned out that when you save an existing experiment, it didn't update the alternatives, so that's what I've added.