Fix caching of winning alternative to prevent recalculation each time the dashboard is loaded
It looks like when this feature was added, the calculation of winning
alternatives was meant to take place only once per day.

The #calc_winning_alternatives method was never called, which was meant
to be saving the experiment's last calc_time. Update the experiment view
to call this method instead of the #estimate_winning_alternative method
directly. Fix caching so that the #calc_time= method is called, rather
than assigning to a local variable. Update calc_time so that number of
days since epoch is stored, rather than the day of month (1-31). Ensure
we're comparing integer values, rather than the string value Redis
returns from #hget.