Rescue from Redis::BaseError in #ab_test
Redis defines a number of different exception classes (see
https://github.com/redis/redis-rb/blob/1bd43346cdf71ba69176612714616289a5a70ed4/lib/redis/errors.rb
for details). When Redis raises an exception, we need to be able to have
the db_failover configuration used, whether it's a failure to connect, a
timeout, or an invalid command.

We ran into this issue when our Redis instance ran out of memory. In
that case (and most others), Split should be able to handle things
gracefully.