DRY Experiment#initialize
Experiment#set_alternatives_and_options appeared both in the `if`
section and in the `else` section. I moved `set_alternatives_and_options`
out of the `if...else` block so that it is more DRY.
I am not sure if setting only options[:alternatives] in the `else`
clause is more readable than explicitly providing a hash with the attributes
to set as it happens in the `if` clause.