CookieAdapter: Fix cookies when @context is an ActionController
Using ab_test from a controller was failing because "cookies" is a
private method of the controller instance which is assigned to @context.

It looks like this was broken in the commit that fixed the multiple
"Set-Cookie" headers bug:
https://github.com/splitrb/split/commit/3523e1c53988d727afb3f28a0208d7984f8fc31b

Using `.send(:cookies)` again when using Rails was the easiest way I
found to fix this.

Note: I was able to add an integration test for the case when it's
called from the controller, but I failed to a regression test that uses
and ActionView context object.